Added subscription to both image topics.
We have two image variables in the computer_vision
class initialisation. We also have a callback method
for each image. The subscribers to the image topics
have been moved to __main__, with the appropriate
callback method passed to them. Each image is then
read and stored in its callback method.